rename run and runall
